73 modify routes to use new authentication api

+++ controllers/poi.controller.js
@@ -1,4 +1,5 @@
 import { POI } from '../models/Poi.js';
+import mongoose from 'mongoose';
 
 export const getPois = async (req, res) => {
     try {
@@ -43,8 +44,8 @@ export const getPoiById = async (req, res) => {
             poi = await POI.findById(req.params.id);
         } else {
             poi = await POI.findOne({
-                _id: req.params.id,
-                user: res.locals.user._id,
+                _id: new mongoose.Types.ObjectId(req.params.id),
+                user: new mongoose.Types.ObjectId(res.locals.user._id),
             });
         }
 
@@ -58,21 +59,29 @@ export const getPoiById = async (req, res) => {
 };
 
 export const deletePoi = async (req, res) => {
+    const user = res.locals.user;
+    const id = req.params.id;
     try {
-        const id = req.params.id;
+        if (user.admin) {
+            const poi = await POI.findById(id);
+            return res.status(200).json({
+                message: 'Poi Deleted successfully',
+            });
+        } else {
+            // check if POI exists and is owned by user
+            const poi = await POI.findOne({
+                _id: id,
+                user: res.locals.user.id,
+            });
 
-        // check if POI exists and is owned by user
-        const poi = await POI.findOne({
-            _id: id,
-            user: res.locals.user._id,
-        });
-        if (!poi) return res.status(404).json({ message: 'POI not found' });
+            if (!poi) return res.status(404).json({ message: 'POI not found' });
 
-        // delete POI
-        await POI.findByIdAndRemove(id);
-        return res.json({ message: 'POI successfully deleted' });
+            // delete POI
+            await POI.findByIdAndRemove(id);
+            return res.json({ message: 'POI successfully deleted' });
+        }
     } catch (error) {
-        return res.status(500).json({ message: 'Internal server error' });
+        return res.status(404).json({ message: `could not delete poi ${id}.` });
     }
 };
 
@@ -103,9 +112,13 @@ export const addPoi = async (req, res) => {
             user: res.locals.user._id,
         });
         await poi.save();
-
-        return res.status(201).json(poi);
+        return res.status(201).json({
+            poi: poi,
+            message: 'Poi Added successfully',
+            redirect: '/index.html',
+        });
     } catch (e) {
+        console.log(e);
         return res.status(500).json({ message: 'Internal server error' });
     }
 };
